Disable compiler checks when outputing metadata
Dropping Nuvoton.

ESP Makefile, updating paths

ESP component.mk path update -- mbedtls, freertos.

Fixing Demo make file

Fixing ESP32 makefile

Fixing ESP demo project

Cleaning make file

Disable cmake build for boards that do not work

Work around permission issue in marvell

Update version string

ESP test project.

amazon-freertos-tests/component.mk is halfway done. commit to get early review

Fixing ESP projects

+++ demos/tcp/aws_tcp_echo_client_single_task.c
@@ -173,7 +173,7 @@ int vStartTCPEchoClientTasks_SingleTasks( bool awsIotMqttMode,
     /* Create the echo client tasks. */
     for( xX = 0; xX < echoNUM_ECHO_CLIENTS; xX++ )
     {
-        snprintf( cNameBuffer, echoMAX_TASK_NAME_LENGTH, "Echo%ld", xX );
+        snprintf( cNameBuffer, echoMAX_TASK_NAME_LENGTH, "Echo%ld", (long int)xX );
         xTaskCreate( prvEchoClientTask,                               /* The function that implements the task. */
                      cNameBuffer,                                     /* Just a text name for the task to aid debugging. */
                      democonfigDEMO_STACKSIZE, /* The stack size is defined in FreeRTOSIPConfig.h. */
